% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_normalise_read_coverage.R
\name{calculate_read_coverage}
\alias{calculate_read_coverage}
\title{Calculate the mean read length for all samples and get the read coverage
across each gene in the input BED file}
\usage{
calculate_read_coverage(
  bam_files_dir,
  bam_files_suffix,
  bed_file,
  gene_level_bed = TRUE,
  count_dups = FALSE
)
}
\arguments{
\item{bam_files_dir}{A character with the full path to the input BAM files}

\item{bam_files_suffix}{A character with the BAM file suffix pattern}

\item{bed_file}{A character with the full path to the input BED file}

\item{gene_level_bed}{A logical used to determine whether the BED file only
contains gene-level features or not (has several lines per gene ID,
accounting for exon, CDS, etc genomic coordinates). If your BED file has
annotation other than gene-level (contains exon, CDS coordinates), set the
value of this parameter to \code{FALSE}.}

\item{count_dups}{A logical used to determine whether to account for
duplicated reads or no when performing read counting. Set this to \code{FALSE}
when determining coverage for variant calling from whole genome sequencing
data.}
}
\value{
A list of two elements: a matrix with the number of reads that
overlap any given region in the BED file for all samples, and a data frame
with the mean read length per sample.
}
\description{
Calculate the mean read length for all samples and get the read coverage
across each gene in the input BED file
}
\examples{
\dontrun{
  res <- calculate_read_coverage(
    bam_files_dir = "data/alignment_outputs",
    bam_files_suffix = "*_recal.bam",
    bed_file = "ref_genomes/PlasmoDB-68_Pfalciparum3D7_genes_only.bed",
    gene_level_bed = TRUE,
    count_dups = FALSE
  )
  per_gene_coverage <- res[["coverage_matrix"]]
  sample_mean_read_length <- res[["per_sample_mean_read_length"]]
}
}
