% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_data_frame.R
\name{choose_files}
\alias{choose_files}
\title{Choose the target images based on the values of the function's arguments}
\usage{
choose_files(images_path, images)
}
\arguments{
\item{images_path}{A character with the full path to the directory where the
images are stored}

\item{images}{A character vector with the path to the images to be converted}
}
\value{
A character vector of the path to the images to be converted
}
\description{
Choose the target images based on the values of the function's arguments
}
\keyword{internal}
