% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{construct_dr_matrix}
\alias{construct_dr_matrix}
\title{Add a drug resistance profile for the specified sample to the data frame of
drug resistance profiles based on potential drug resistance mutations}
\usage{
construct_dr_matrix(dat, sample_name, tmp_others_dr)
}
\arguments{
\item{dat}{A data frame with the resistance profile of the target sample
based on the potential drug resistance mutations}

\item{sample_name}{A character with a sample name}

\item{tmp_others_dr}{A data frame used to store the sample profiles}
}
\value{
A data frame potential drug resistance profile augmented with one
column corresponding the profile of the target sample
}
\description{
Add a drug resistance profile for the specified sample to the data frame of
drug resistance profiles based on potential drug resistance mutations
}
\keyword{internal}
