% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{find_causative_snp}
\alias{find_causative_snp}
\title{Scan all SNPs in codon and find which causes ref_aa → alt_aa}
\usage{
find_causative_snp(codon_bases, codon_seq, ref_aa, alt_aa, strand)
}
\arguments{
\item{codon_bases}{A list with the codon coordinates of the protein mutation
on each exon of the target gene}

\item{codon_seq}{A character with the reference codon sequence}

\item{ref_aa}{A character with the provided reference amino acid of the
protein change of interest}

\item{alt_aa}{A character with the provided alternative amino acid of the
protein change of interest}

\item{strand}{A character vector with strand of each exon of the gene}
}
\value{
A list with
}
\description{
Scan all SNPs in codon and find which causes ref_aa → alt_aa
}
\keyword{internal}
