% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{get_cds}
\alias{get_cds}
\title{Extract CDS intervals for a given PlasmoDB gene ID}
\usage{
get_cds(gff_df, gene_id)
}
\arguments{
\item{gff_df}{A data frame generated by the \code{parse_gff()} function}

\item{gene_id}{A character with gene ID}
}
\value{
A data frame with the coding sequence information for the specified
gene ID. It returns \code{NULL} when no match is found.
}
\description{
Extract CDS intervals for a given PlasmoDB gene ID
}
\keyword{internal}
