% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_chrom}
\alias{get_chrom}
\title{Extract chromosome names from coverage matrix row names}
\usage{
get_chrom(x)
}
\arguments{
\item{x}{A vector of the row names of the coverage matrix}
}
\value{
A character vector with the chromosome names only
}
\description{
Extract chromosome names from coverage matrix row names
}
\keyword{internal}
