% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{get_codon_positions}
\alias{get_codon_positions}
\title{Map a protein position to 3 genomic coordinates (the codon)}
\usage{
get_codon_positions(cds_df, protein_pos)
}
\arguments{
\item{cds_df}{A data frame generated by the \code{get_cds()} function}

\item{protein_pos}{An integer with the position of the protein mutation}
}
\value{
A list with the corresponding genomic coordinates (chromosome,
position, and strand) of the three nucleotides that form the codon of the
given protein position. It returns \code{NULL} when no match is found.
}
\description{
Map a protein position to 3 genomic coordinates (the codon)
}
\keyword{internal}
