% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{get_confirmed_dr_variants}
\alias{get_confirmed_dr_variants}
\title{Summarise the malaria-profiler outputs for confirmed drug resistance
mutations}
\usage{
get_confirmed_dr_variants(json, variants, sample_name)
}
\arguments{
\item{json}{A list object obtained from reading an output file in JSon format}

\item{variants}{A data frame that will be used to store the variant profile
for each sample}

\item{sample_name}{A character with a sample name}
}
\value{
An updated variant table where the profile of the specified sample
is added
}
\description{
Summarise the malaria-profiler outputs for confirmed drug resistance
mutations
}
\keyword{internal}
