% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_coverage_stats}
\alias{get_coverage_stats}
\title{Get read coverage statistics per sample and per region}
\usage{
get_coverage_stats(coverage_matrix, mean_read_length)
}
\arguments{
\item{mean_read_length}{A data frame with the mean read length per sample.
This is the first element of the output from the
\code{calculate_read_coverage()}
function.}

\item{@param}{coverage_matrix A numeric matrix with the coverage data. This
is the first element of the output from the \code{calculate_read_coverage()}
function.}
}
\value{
A list of three data frames: one with the total, mean coverage, and
the mean read length per sample. Another one with the region's genomic
coordinates, lengths, and median number of read across all samples. The
third one corresponds to the read coverage matrix.
}
\description{
The function returns: the total number of good quality reads (accounted for
during variant calling process) per sample across all genes, the mean number
of good reads per sample across all genes, the gene length, the median number
of reads covered by each gene, and mean read length per sample.
}
\examples{
\dontrun{
  cov_stats <- get_coverage_stats(
    coverage_matrix = per_gene_coverage,
    mean_read_length = sample_mean_read_length
  )
  coverage_matrix <- cov_stats[["coverage_matrix"]]
  samples_cov_stats <- cov_stats[["meta"]]
  regions_cov_stats <- cov_stats[["details"]]
}
}
