% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{get_drug}
\alias{get_drug}
\title{Extract drug name from variants file}
\usage{
get_drug(x)
}
\arguments{
\item{x}{A character that corresponds to an entry of the Info column in the
variant file}
}
\value{
A character with the drug name
}
\description{
Extract drug name from variants file
}
\keyword{internal}
