% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_end_pos}
\alias{get_end_pos}
\title{Extract region's end position from a vector of genomic coordinates}
\usage{
get_end_pos(x)
}
\arguments{
\item{x}{A vector of the genomic coordinates}
}
\value{
A numeric vector with the region's end positions only
}
\description{
Extract region's end position from a vector of genomic coordinates
}
\keyword{internal}
