% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_gene_ids}
\alias{get_gene_ids}
\title{Extract gene IDs from coverage matrix row names}
\usage{
get_gene_ids(x)
}
\arguments{
\item{x}{A vector of the row names of the coverage matrix}
}
\value{
A character vector with the genes IDs only
}
\description{
Extract gene IDs from coverage matrix row names
}
\keyword{internal}
