% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_genomic_coordinates}
\alias{get_genomic_coordinates}
\title{Extract genes genomic coordinates from the row names of the read coverage
matrix}
\usage{
get_genomic_coordinates(coverage_matrix)
}
\value{
A list of two elements: the coverage data matrix where the row names
are only composed of the gene IDs, and a data frame with the gene's
genomic coordinates.
}
\description{
Extract genes genomic coordinates from the row names of the read coverage
matrix
}
\keyword{internal}
