% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{get_inferred_region}
\alias{get_inferred_region}
\title{Extract the prediction region of origin for the specified sample}
\usage{
get_inferred_region(json, metadata, region_proba, sample_name)
}
\arguments{
\item{json}{A list object obtained from reading an output file in JSon format}

\item{region_proba}{A data frame used to store the prediction probabilities
for each region}

\item{sample_name}{A character with a sample name}
}
\value{
A list with the updated sample metadata and region probabilities
data frame
}
\description{
Extract the prediction region of origin for the specified sample
}
\keyword{internal}
