% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{get_other_dr_variants}
\alias{get_other_dr_variants}
\title{Get the sample profiles for the potential drug resistance mutations}
\usage{
get_other_dr_variants(nested_list, sample_name, tmp_others_dr)
}
\arguments{
\item{nested_list}{A list with the information on the potential drug
resistance mutations found on the given sample}

\item{sample_name}{A character with a sample name}

\item{tmp_others_dr}{A data frame used to store the sample profiles}
}
\value{
A data frame with the sample profiles augmented with a new column
corresponding to the profile of the specified sample
}
\description{
Get the sample profiles for the potential drug resistance mutations
}
\keyword{internal}
