% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_pos}
\alias{get_pos}
\title{Extract region's genomic coordinates from coverage matrix row names}
\usage{
get_pos(x)
}
\arguments{
\item{x}{A vector of the row names of the coverage matrix}
}
\value{
A character vector with the region's genomic coordinates only
}
\description{
Extract region's genomic coordinates from coverage matrix row names
}
\keyword{internal}
