% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{get_ref_codon}
\alias{get_ref_codon}
\title{Extract reference codon sequence from genome FASTA}
\usage{
get_ref_codon(genome, codon_bases, strand)
}
\arguments{
\item{genome}{A DNAString object with the reference genome sequence}

\item{codon_bases}{A list with the codon coordinates of the protein mutation
on each exon of the target gene}

\item{strand}{A character vector with strand of each exon of the gene}
}
\value{
A character with the reference codon obtained from the reference
genome sequence and codon coordinates
}
\description{
Extract reference codon sequence from genome FASTA
}
\keyword{internal}
