% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_read_coverage_statistics.R
\name{get_start_pos}
\alias{get_start_pos}
\title{Extract region's start position from a vector of genomic coordinates}
\usage{
get_start_pos(x)
}
\arguments{
\item{x}{A vector of the genomic coordinates}
}
\value{
A numeric vector with the region's start positions only
}
\description{
Extract region's start position from a vector of genomic coordinates
}
\keyword{internal}
