% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variant_coverage.R
\name{get_variant_coverage}
\alias{get_variant_coverage}
\title{Check whether a VCF file contains SNPs overlapping a set of genes of interest}
\usage{
get_variant_coverage(vcf_file, genes_file, annotation_file)
}
\arguments{
\item{vcf_file}{A character with the path to the input \code{VCF} file}

\item{genes_file}{A character with the path to the \code{BED} file with the
genomic coordinates of the genes of interest. It must contain at least the
CHROM, START, and END columns}

\item{annotation_file}{A character with the path to file with the annotation
columns for every variant}
}
\value{
Writes out files with the variants overlapping each gene and the
gene's variant counts.
}
\description{
Check whether a VCF file contains SNPs overlapping a set of genes of interest
}
\examples{
\dontrun{
  get_variant_coverage(
    vcf_file = file.path("data", "variant_calling_outputs",
                         "biallelic_snps.vcf.gz"),
    genes_file = file.path("ref_genomes", "genes_cordinates_on_v68.txt"),
    annotation_file = file.path("data", "variant_calling_outputs",
                                "annotation_fields.txt")
  )
}
}
