% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{import_from_csv}
\alias{import_from_csv}
\title{Generate variant data frame from a CSV variant file}
\usage{
import_from_csv(variants_file)
}
\arguments{
\item{variants_file}{A character with the path to the variant file}
}
\value{
A data frame of three columns with the gene IDs, the mutation, and
the name of the drug
}
\description{
Generate variant data frame from a CSV variant file
}
\keyword{internal}
