% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome.R
\name{map_protein_change_to_genome}
\alias{map_protein_change_to_genome}
\title{Map Plasmodium falciparum protein change to a Single Nucleotide Polymorphism
(SNP)}
\usage{
map_protein_change_to_genome(
  ref_genome_file,
  gff_file,
  gene_name,
  protein_position,
  ref_aa,
  alt_aa
)
}
\arguments{
\item{ref_genome_file}{A character with the path to the reference genome FASTA
file}

\item{gff_file}{A character with the path the to reference genome GFF file}

\item{gene_name}{A character with the target gene name or gene ID}

\item{protein_position}{An integer with the protein change position}

\item{ref_aa}{A character with the one-letter amino acid code for the
reference amino acid}

\item{alt_aa}{A character with the one-letter amino acid code for the
alternative amino acid}
}
\value{
Prints a table with the codon information for all possible SNPs
which can cause the given protein change
}
\description{
Map Plasmodium falciparum protein change to a Single Nucleotide Polymorphism
(SNP)
}
\examples{
\dontrun{
  snp <- map_protein_change_to_genome(
    ref_genome_file = "PlasmoDB-68_Pfalciparum3D7_Genome.fasta",
    gff_file = "PlasmoDB-68_Pfalciparum3D7.gff",
    gene_name = "PF3D7_0523000",
    protein_position = 86,
    ref_aa = "N",
    alt_aa = "Y"
  )
}
}
