% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_and_normalise_read_coverage.R
\name{normalize_coverage_matrix}
\alias{normalize_coverage_matrix}
\title{Normalise coverage matrix}
\usage{
normalize_coverage_matrix(coverage_matrix)
}
\arguments{
\item{coverage_matrix}{A numeric matrix with the read count per region across
all samples. This is the output of the \code{calculate_read_coverage()}
function.}
}
\value{
A numeric matrix obtained from the normalization process explained
in the function description.
}
\description{
The function performs relative read count abundance normalization i.e. for
each sample, divides a region read count by the sum read count across all
regions (also known as total sum scaling). This gives the relative coverage
of each gene as a proportion of total coverage per sample.
Then computes the z-scores on the normalised coverage data.
}
\examples{
\dontrun{
  res <- calculate_read_coverage(
    bam_files_dir = "data/alignment_outputs",
    bam_files_suffix = "*_recal.bam",
    bed_file = "ref_genomes/PlasmoDB-68_Pfalciparum3D7_genes_only.bed",
    gene_level_bed = TRUE,
    count_dups = FALSE
  )
  per_gene_coverage <- res[["coverage_matrix"]]
  res <- normalize_coverage_matrix(per_gene_coverage)
}
}
