% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{parse_gff}
\alias{parse_gff}
\title{Parse GFF file into a data frame}
\usage{
parse_gff(gff_file)
}
\value{
A data frame with the gene genomic coordinates and feature
annotation
}
\description{
Parse GFF file into a data frame
}
\keyword{internal}
