% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summurise_fastqscreen.R
\name{parse_output_file}
\alias{parse_output_file}
\title{Parse Fastq-Screen output file and check coverage of the target genome}
\usage{
parse_output_file(output_file)
}
\arguments{
\item{output_file}{A character with the path to the Fastq-Screen output file}
}
\value{
A list of two elements: A numeric with number of reads that do not
map to any of the specified genome, and a data frame with screening result
}
\description{
Parse Fastq-Screen output file and check coverage of the target genome
}
\keyword{internal}
