% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{plasmor-package}
\alias{plasmor}
\alias{plasmor-package}
\title{plasmor: Analyse Plasmodium falciparum data}
\description{
The package is a library of functions to analyse different types of data generated by a researcher working in the field of parasitology, specially on the malaria parasite P. falciparum.
}
\author{
\strong{Maintainer}: Karim Mané \email{karimanee@outlook.com} (\href{https://orcid.org/0000-0002-9892-2999}{ORCID}) [copyright holder]

}
\keyword{internal}
