% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summurise_fastqscreen.R
\name{plot_sample_screening}
\alias{plot_sample_screening}
\title{Plot the Fastq-Screen output for a specific sample}
\usage{
plot_sample_screening(dir, sample)
}
\arguments{
\item{dir}{A character with the path to the folder where the Fastq-Screen
files are stored}

\item{sample}{A character with the name of a target sample}
}
\value{
Invisibly returns the screening output statistics for the specified
sample
}
\description{
Plot the Fastq-Screen output for a specific sample
}
\examples{
\dontrun{
  plot_sample_screening(
    dir = "/Volumes/NO NAME/fastq_screen_outputs",
    sample = "29504"
  )
}
}
