% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs-helpers.R
\name{read_confirmed_variants}
\alias{read_confirmed_variants}
\title{Import variants file depending on whether it's in a \code{.csv} or \code{.json} format}
\usage{
read_confirmed_variants(variants_file)
}
\arguments{
\item{variants_file}{A character with the path to the variant file}
}
\value{
A data frame of three columns with the gene IDs, the mutation, and
the name of the drug
}
\description{
Import variants file depending on whether it's in a \code{.csv} or \code{.json} format
}
\keyword{internal}
