% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_protein_change_to_genome-helpers.R
\name{resolve_gene_id}
\alias{resolve_gene_id}
\title{Resolve gene name to PlasmoDB ID}
\usage{
resolve_gene_id(gene_name, gene_lookup)
}
\arguments{
\item{gene_name}{A character with the gene name}

\item{gene_lookup}{A named vector of the predefined set of resistance genes}
}
\value{
The corresponding gene ID for the provided gene name. The function
raises an error if the provided gene name is not part of the set of
predefined target genes.
}
\description{
Resolve gene name to PlasmoDB ID
}
\keyword{internal}
