% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_profiler_outputs.R
\name{summarise_mp_outputs}
\alias{summarise_mp_outputs}
\title{Summarise \code{malaria-profiler} outputs}
\usage{
summarise_mp_outputs(
  target_dir,
  variants_file,
  metadata_file,
  mutation_type = c("confirmed", "potential"),
  geo_localisation = TRUE
)
}
\arguments{
\item{target_dir}{A character with the full path to the folder where the
output files (.json files) from malaria-profiler are stored}

\item{variants_file}{A character with the path to the file with the confirmed
mutations associated to the resistance to the drugs of interest}

\item{metadata_file}{A character with the full path to file with the sample
metadata}

\item{mutation_type}{A character that determines whether to summarise the
outputs for the confirmed or potential drug resistance mutations. Possible
values are \code{confirmed} or \code{potential}. Default is \code{confirmed}}

\item{geo_localisation}{A logical used to determine whether to summarise the
outputs about the potential region of origin of the samples. If \code{TRUE}, a
column with inferred sample potential regions of origin will be added to
the sample metadata table. Default is \code{TRUE}}
}
\value{
A list with the following three elements:
\enumerate{
\item variants: a data frame with the genomic coordinates of the
detected drug resistance mutations and \code{x} columns of \code{0} and \code{1}
used to determine whether the sample carries a particular mutation
(\code{1}) or no (\code{0}), where \code{x} is the number of processed samples.
\item metadata: a data frame with the metadata table augmented with two
columns: the sample origins, and the fraction of detected
SNP markers used for geographic origin prediction
\item region_probabilities: a data frame with the probabilities obtained
during the geographic origin prediction. This is empty when
\code{geo_localisation = FALSE}
}
}
\description{
Summarise \code{malaria-profiler} outputs
}
\examples{
\dontrun{
  # establish the connection to the database
  results <- summarise_mp_outputs(
    target_dir = tempdir(),
    variants_file = file.path(tempdir(), "variant.csv"),
    metadata_file = file.path(tempdir(), "metadata.txt"),
    mutation_type = "confirmed",
    geo_localisation = "TRUE"
  )
}
}
