% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summurise_fastqscreen.R
\name{summarise_screening_outputs}
\alias{summarise_screening_outputs}
\title{Parse Fastq-Screen output file to extract genome mapping statistics.}
\usage{
summarise_screening_outputs(dir, target_genome)
}
\arguments{
\item{dir}{A character with the path to the folder where the Fastq-Screen
files are stored}

\item{target_genome}{A character with the name of the target genome as
specified in the Fastq-Screen config file}
}
\value{
A data frame with seven columns when there are samples with a
coverage is less than 50\% or the percent of no hits is greater than 5\%
}
\description{
If the coverage of the target genome is less than 50\% or the percent of
reads mapping to none of the genomes is greater than 5\%, a comment is
associated to that read.
Also, the \%reads mapping to the Human genome is reported. Files where there
are reads mapping uniquely to other Plasmodium genome will have the name of
these organisms in the 'others' column.
}
\examples{
\dontrun{
  screening_summary <- summarise_screening_outputs(
    dir = "/Volumes/NO NAME/fastq_screen_outputs",
    target_genome = "falciparum"
  )
}
}
