% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genome.R
\name{update_genome}
\alias{update_genome}
\title{Substitute a nucleotide on the reference genome Fasta file}
\usage{
update_genome(
  ref_genome_file,
  chrom,
  pos,
  new_base,
  expected_base = NULL,
  output_dir = NULL
)
}
\arguments{
\item{ref_genome_file}{A character with the path to the reference genome FASTA
file}

\item{chrom}{A character with name of the chromosome which sequence will be
updated}

\item{pos}{A numeric with the position of the nucleotide to be substituted on
the chromosome of interest}

\item{new_base}{A character with the new nucleotide to introduced at the
given position}

\item{expected_base}{A character with the reference nucleotide at the given
position. Default is \code{NULL}.}
}
\value{
A message specifying the path to the folder where the updated genome
is stored
}
\description{
Substitute a nucleotide on the reference genome Fasta file
}
\examples{
\dontrun{
  update_genome(
    ref_genome_file = "PlasmoDB-68_Pfalciparum3D7_Genome.fasta",
    chrom = "Pf3D7_08_v3",
    pos = 549685,
    new_base = "G",
    expected_base = "G",
    output_dir = NULL
  )
}
}
