% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_data_frame.R
\name{convert_to_png}
\alias{convert_to_png}
\title{Convert images to PNG format}
\usage{
convert_to_png(images = NULL, images_path = NULL)
}
\arguments{
\item{images}{A character vector with the path to the images to be converted}

\item{images_path}{A character with the full path to the directory where the
images are stored}
}
\value{
Invisibly return a vector of the converted images
}
\description{
Convert images to PNG format
}
\examples{
# convert a specific image
convert_to_png(
  images <- system.file("images", "Image003.heic", package = "plasmor")
)
}
